Refactor, to avoid problem where jest not finding tests
